Refactor builder; Move into smaller files before further refactoring
