+++ Sources/libgir2swift/emitting/Documentation/gtk2swiftdoc.swift
