+++ Sources/libgir2swift/emitting/FreeFunctions/emit-functions.swift
@@ -1,0 +1,8 @@
+import Foundation
+
+/// Swift code representation of a free standing function
+public func swiftCode(_ f: GIR.Function) -> String {
+    let code = functionCode(f)
+    return code
+}
+
