+++ Sources/libgir2swift/emitting/supporting/c2swift.swift
