+++ Sources/libgir2swift/emitting/supporting/girtypes+swift.swift
